couch2pg now understands edits and deletes medic/medic-webapp#2323
Large refactor to add support for understanding edits and deletes from
couchdb. We now use the PouchDB js library and use the changes feed
so we can detect edits and deletes.

This was a reasonably fundamental change and so there were a lot of
changes to the code, tests etc.