Remove XML parsing medic/medic-webapp#2432
Removes all XML parsing. Because there is no XML parsing we no longer
need to pass the changes from couch2pg (and couch2pg no longer needs
to generate them, but we do it anyway for now). Because we no longer
need to pass changes we don't need to worry about losing data, so we
run importAll() now instead.

All xmlforms does now is make sure the helper SQL exists (views and
the like) and that at the end of each run any materialised views that
exist are updated.